requirements.txt
<<<<<<< HEAD
python-dotenv~=0.21.1
requests~=2.28.2
python-telegram-bot=13.15
=======
Django==4.1.7
>>>>>>> origin/DB_Creation
